requirements.txt
torch~=2.8.0
boto3~=1.40.39
requests~=2.32.5
botocore~=1.40.39
tqdm~=4.67.1
numpy~=2.3.3
pillow~=11.3.0
transformers~=4.44.0
matplotlib~=3.10.7
seaborn~=0.13.2
pandas~=2.3.3
psutil~=7.1.2
gputil~=1.4.0
bert-score~=0.3.13